Fix issue where disabling of I/O was checking the I/O control registers when the disabled bit is on the memory control register. Tidy up VDP control word management Fixed issue where when using paging, it was limiting to ushort so could not access more than 3 pages of ROM memory.  ushort makes sense since addresses are only 16bit but since we don't have any paging hardware and not emulating swapping pages, when we read cartridge ROM we want to the full ROM address space which needs 32bits.  This meant when games such a Sonic attempted to page in ROM memory higher than 3, it would end up accessing the first 32kb of memory.  This causes graphics to write garbage in larger ROMS such as Sonic but meant smaller test/homebrew ROMS worked properly since never needed access to the higher sections of ROM memory.

+++ Kmse.Console/Logging/SerilogMemoryLogger.cs
@@ -42,7 +42,7 @@ public class SerilogMemoryLogger : IMemoryLogger
         _logger.Debug("Read system RAM at address {Address:X4}, got {Data}", address, data);
     }
 
-    public void CartridgeRead(ushort address, byte data)
+    public void CartridgeRead(int address, byte data)
     {
         if (!_verboseLoggingEnabled)
         {
